[FIX] web: navbar menu not shown on tablet
Before this fix, 'currentApp' returned 'undefined' on mobile but
there is no reason not to return the current app on mobile.
This has been fixed in enterprise repository.

Now that 'currentApp' returns a value, we have to explicitly hide
'menuItem' for small screen.

Steps to reproduce:
* Open Odoo on tablet
* Select an App => BUG

opw-2687424
opw-2688832

closes odoo/odoo#80258

Related: odoo/enterprise#22478
Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>