[FIX] web_editor: prevent crash when going to unlisted editor file
When switching to files that can only appear in debug mode in the HTML /
SCSS editor and editing them, the page reloads... and crash for no
reason. This was because the filter that allows those files to appear
is lost after reloading, now we simply ignore the file switch if not
present in the default filter.

task-2698641

X-original-commit: f43b514f10078e925e2ccfcfac6dc09e6961bca0
Part-of: odoo/odoo#80497