[FIX] account_journal: ignore line_section and line_notes in account_control
When using account control on a journal, line_section and line_note do not
block the process even though they are not in the allowed accounts (they
don't have an account at all).
But when adding new accounts in the account_control_ids field, the
constraint is triggered by those lines.
This PR makes the constraint ignore those lines when assessing if
there's an issue.

opw-2677597

closes odoo/odoo#80306

X-original-commit: 9768c6e83ba906bf392095794be98a90e4e4a6ef
Signed-off-by: Florian Gilbert <flg@odoo.com>