[FIX] web: fix chineese calendar format
Issue:
The way we generate the dayNamesMin is not compatible with the chineese
translation provided by "moment.js".
This is because the dayNamesShort contains an extra character in front
週 (week). Because of that the names of the days on the
small calendar are all the character 週 (week).

Solution:
There is a variable (weekdaysMin) which is unused and contain a shorter
version of the days this variable is consitent in the other languages so
it should have no side effect and will fix the chinese translation as it
only contains one character.

Introduced by: 5f30a4cd2ac6d2c659fb06bc2fba31c8956e7e91

task-2692910

closes odoo/odoo#80445

X-original-commit: f7e8016e07bd098acf78de89719198cf972a3f5d
Signed-off-by: Samuel Degueldre <sad@odoo.com>
Signed-off-by: fja-odoo <fja@odoo.com>