[FIX] web: allow navigation by changing the value of browser.location
The browser object is meant to allow easily patching some native browser
behaviour when writing tests. Outside of test code, it should behave
like the corresponding properties on the window object.

Previously, the location property of the browser object held a reference
to the native window location property. While this works perfectly fine
for reading or writing things on the location object, a common use case
is to overwrite window.location with a string to navigate to a url. This
use case was not covered by browser.location

This commit fixes that by replacing the browser.location property with a
getter/setter pair, where setting a value on browser.location will do
the same thing on window.location

closes odoo/odoo#80374

Related: odoo/enterprise#22576
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>