[FIX] web_editor: fix illustration crop issue
when we crop illustrations using the editor crop tool,
the resulted image will disappear as soon as we save the update.

we have this issue because we cannot get the right svg string using
toDataURL on the new cropped canvas with mimetype = 'image/svg+xml'
(see 'web_editor.image_processing' > 'applyModifications')

To solve this issue we set svg illustrations as uncroppable (for now).

task-2312878

closes odoo/odoo#81168

X-original-commit: a1e1cb8399c0f8f9d2911091a45cca0a7fb29744
Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>