[FIX] pos_cache: error when loading missing products
As part of loading orders, we need to make sure we load the missing
products. This process relies on the existence of the `product.product`
model declaration in the PosModel. However, since we are loading
products differently here in pos_cache, the `product.product` model
is removed from the list of models.

We are now return the `product.product` model to the list of models
after loading.

closes odoo/odoo#80780

Signed-off-by: Trinh Jacky (trj) <trj@odoo.com>