[FIX] mail: properly render following canceled patching
Before this commit, sometimes component do not properly
render changes in the models. Although the modeling is correct,
there is a problem in the component not re-rendering itself.

The shouldUpdateBasedOnProps component hook did not consider
update attempts that could be canceled. This commit adds this
consideration, so that it ensures component is eventually
rendered correctly

Task-2410314
Task-2705078

closes odoo/odoo#80863

X-original-commit: 1ce2f9250eba503be885218d2c1b580db7469a4c
Signed-off-by: Sébastien Theys (seb) <seb@odoo.com>