[FIX] uom: Correct uom creation link
Since 8c1193813815b204e8b168d083956f725ad4df15 UoM should be created through the UoM category tab as the UoM form view is now unusable.
To ensure that this is the case, we remove menu link to the form view and update the other to lead to UoM category instead.

opw-2702953

closes odoo/odoo#80659

Signed-off-by: Damhaut Florian (flda) <flda@odoo.com>