[FIX] website: fix inconsistent edit menu test
[1] introduced a new test on the EditMenuDialog. This test used an
unreliable event 'unload' as a workaround to go to the next step when
saving the dialog.

This commit removes the use of this event, resulting in crashing
randomly the test, to use the .o_wait_reload class, so that the tour
correctly waits for the page to be reloaded.

task-2513588

[1]: 301d8628be8e287e56e67ac2385ea7388d341fb7

closes odoo/odoo#81094

Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>